requirements.txt
flask
flask-cors
pygame
elevenlabs
python-dotenv